Started refactoring to global eval pipe
